+++ src/Store.cpp
@@ -199,7 +199,7 @@ std::vector<std::pair<ULONG, LPMAPINAMEID>> Store::lookupPropIdInputs(
 			auto name = convert::utf8::to_utf16(*id.named->name);
 
 			namedId->ulKind = MNID_STRING;
-			CORt(::MAPIAllocateMore((name.size() + 1) * sizeof(wchar_t),
+			CORt(::MAPIAllocateMore(static_cast<ULONG>((name.size() + 1) * sizeof(wchar_t)),
 				namedId.get(),
 				reinterpret_cast<void**>(&namedId->Kind.lpwstrName)));
 			CFRt(namedId->Kind.lpwstrName != nullptr);
@@ -308,7 +308,7 @@ std::vector<std::pair<ULONG, LPMAPINAMEID>> Store::lookupPropIds(const std::vect
 	{
 		mapi_ptr<SPropTagArray> propIds;
 
-		CORt(::MAPIAllocateBuffer(CbNewSPropTagArray(resolve.size()),
+		CORt(::MAPIAllocateBuffer(CbNewSPropTagArray(static_cast<ULONG>(resolve.size())),
 			reinterpret_cast<void**>(&out_ptr { propIds })));
 		CFRt(propIds != nullptr);
 
@@ -347,7 +347,7 @@ std::vector<std::pair<ULONG, LPMAPINAMEID>> Store::lookupPropIds(const std::vect
 				std::wstring_view source { name->Kind.lpwstrName };
 
 				namedId->ulKind = MNID_STRING;
-				CORt(::MAPIAllocateMore((source.size() + 1) * sizeof(wchar_t),
+				CORt(::MAPIAllocateMore(static_cast<ULONG>((source.size() + 1) * sizeof(wchar_t)),
 					namedId.get(),
 					reinterpret_cast<void**>(&namedId->Kind.lpwstrName)));
 				CFRt(namedId->Kind.lpwstrName != nullptr);
@@ -492,7 +492,7 @@ std::vector<std::shared_ptr<object::Property>> Store::GetProperties(
 
 		mapi_ptr<SPropTagArray> propIds;
 
-		CORt(::MAPIAllocateBuffer(CbNewSPropTagArray(resolved.size()),
+		CORt(::MAPIAllocateBuffer(CbNewSPropTagArray(static_cast<ULONG>(resolved.size())),
 			reinterpret_cast<void**>(&out_ptr { propIds })));
 		CFRt(propIds != nullptr);
 		propIds->cValues = static_cast<ULONG>(resolved.size());
@@ -979,7 +979,7 @@ void Store::LoadRootFolders(service::Directives&& fieldDirectives)
 	constexpr auto c_folderSorts = Folder::GetFolderSorts();
 	mapi_ptr<SSortOrderSet> folderSorts;
 
-	CORt(::MAPIAllocateBuffer(CbNewSSortOrderSet(c_folderSorts.size()),
+	CORt(::MAPIAllocateBuffer(CbNewSSortOrderSet(static_cast<ULONG>(c_folderSorts.size())),
 		reinterpret_cast<void**>(&out_ptr { folderSorts })));
 	CFRt(folderSorts != nullptr);
 	folderSorts->cSorts = static_cast<ULONG>(c_folderSorts.size());
@@ -1043,7 +1043,7 @@ mapi_ptr<SPropTagArray> Store::GetFolderProperties() const
 	constexpr auto c_folderProps = Folder::GetFolderColumns();
 	mapi_ptr<SPropTagArray> folderProps;
 
-	CORt(::MAPIAllocateBuffer(CbNewSPropTagArray(c_folderProps.size()),
+	CORt(::MAPIAllocateBuffer(CbNewSPropTagArray(static_cast<ULONG>(c_folderProps.size())),
 		reinterpret_cast<void**>(&out_ptr { folderProps })));
 	CFRt(folderProps != nullptr);
 	folderProps->cValues = static_cast<ULONG>(c_folderProps.size());
@@ -1057,7 +1057,7 @@ mapi_ptr<SPropTagArray> Store::GetItemProperties() const
 	constexpr auto c_itemProps = Item::GetItemColumns();
 	mapi_ptr<SPropTagArray> itemProps;
 
-	CORt(::MAPIAllocateBuffer(CbNewSPropTagArray(c_itemProps.size()),
+	CORt(::MAPIAllocateBuffer(CbNewSPropTagArray(static_cast<ULONG>(c_itemProps.size())),
 		reinterpret_cast<void**>(&out_ptr { itemProps })));
 	CFRt(itemProps != nullptr);
 	itemProps->cValues = static_cast<ULONG>(c_itemProps.size());
